% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC2.R
\name{MAC2}
\alias{MAC2}
\title{MAC}
\usage{
MAC2(x, y, k = 20)
}
\arguments{
\item{x}{A list of object.}

\item{y}{Another list of object.}

\item{k}{The sub-sampling size of MAC.}
}
\value{
The statistic value of MAC.
}
\description{
Maximum adjusted chi-squared statistics for testing the difference between two objects.
}
