% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBET.R
\name{RBET}
\alias{RBET}
\title{RBET.}
\usage{
RBET(obj, rgene, batch, k = 50)
}
\arguments{
\item{obj}{A Seurat object of single cell data.}

\item{rgene}{A vector of reference genes.}

\item{batch}{The column name for batch information in obj.}

\item{k}{The sub-sampling size for RBET.}
}
\value{
The batch effect of dataset.
}
\description{
Reference-based batch effect testing.
}
