% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_sel_data.R
\name{rg_sel_data}
\alias{rg_sel_data}
\title{Data-based reference genes selection.}
\usage{
rg_sel_data(obj, batch, dim = 20, resolution = 0.2, p = 0.05, k = 30)
}
\arguments{
\item{obj}{A Seurat object with batch information.}

\item{batch}{The column name for batch information in obj.}

\item{dim}{The number of PCA used in RunUMAP and FindNeighbors.}

\item{resolution}{The resolution of clusters in FindClusters.}

\item{p}{The top p genes are retained when selecting invariable genes in clusters and across clusters, with p ranging from 0 to 1.}

\item{k}{The top k genes are retained when selecting the most differently expressed genes across batches.}
}
\value{
A vector of reference genes.
}
\description{
Automatically select reference genes from data, based on the extent to which they are stably expressed in batches and differently expressed across batches.
}
