% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_sel_literature.R
\name{rg_sel_literature}
\alias{rg_sel_literature}
\title{Literature-based reference genes selection.}
\usage{
rg_sel_literature(obj, batch, celltype, rgene)
}
\arguments{
\item{obj}{A Seurat object with batch information.}

\item{batch}{The column name for batch information in obj.}

\item{celltype}{The column name for cell type information in obj.}

\item{rgene}{Housekeeping genes validated by previous literature.}
}
\value{
A vector of reference genes.
}
\description{
Select reference genes based on the extent to which they are differently expressed across batches.
}
